Recon2obj now parses and correctly applies transforms of Reconstruct contours
Add code to handle forward and inverse coordinate system transforms of contours
from Reconstruct files.  Code was adapted from nform.cpp in the repository for synapseweb/Reconstruct.

The transform coefficients in Reconstruct XML files are the coefficients of the
transform polynomials that go from World coordinates to contour coordinates. So
to transform from contour coordinates to World coordinates one must apply the
inverse of the transform polynomials.  The functions in reconstruct_transform.c
are adapted from the Reconstruct source code (nform.cpp) to compute the forward
and inverse transform.